Lots of changes, mostly around surfaces
There was somewhat of a turnaround, where instead of trying to present
a simple interface to SDL2, I started moving towards a more inclusive
set of of drawing structs. In SDL2, a Texture and a Surface are two
different pieces of functionality, in this library, I want to merge the
two into a Surface struct, which is kind of a super-Surface, that also
acts like a simple texture when that's all you want. As a Surface can
have 1 Texture per Window, I'll need to provide a Window-specific
Surface for people who have multiple Textures and want to modify
textures in different ways for different Windows. Probably the easiest
way it to have a Surface.Fork() function.

Also started on a games subdirectory which will showcase complete
applications (games) written with this library. Each game includes
things like Menus, Gameplay, Saves, etc.

Also new is the gui subdirectory, which will eventually do a GUI
interface with click handling and hopefully some degree of UI
generation. Still playing with how this is going to work, but it sort of
works now.